Crate compositing_traits

The interface to the compositing crate.

Modules

display_list

Defines data structures which are consumed by Paint.

largest_contentful_paint_candidate

Definitions for Largest Contentful Paint Candidate and Largest Contentful Paint.

rendering_context

viewport_description

This module contains helpers for Viewport

Structs

CompositionPipeline

The subset of the pipeline that is needed for layer composition.

CrossProcessPaintApi

A mechanism to send messages from ScriptThread to the parent process’ WebRender instance.

PaintProxy

Sends messages to Paint.

PainterSurfmanDetails

PainterSurfmanDetailsMap

PipelineExitSource

What entity is reporting that a Pipeline has exited. Only when all have done this will the renderer discard its details.

SendableFrameTree

WebRenderExternalImageHandlers

WebRender External Image Handler implementation.

WebRenderExternalImageIdManager

WebRenderExternalImageIdManagerInner 🔒

List of WebRender external images to be shared among all external image consumers (WebGL, Media, WebGPU). It ensures that external image identifiers are unique.

Enums

ExternalImageSource

The source for an external image.

ImageUpdate

Serializable image updates that must be performed by WebRender.

PaintMessage

Messages from (or via) the constellation thread to Paint.

SerializableImageData

Serialized ImageData. It contains IPC byte channel receiver to prevent from loading bytes too slow.

WebRenderImageHandlerType

Type of WebRender External Image Handler.

Traits

WebRenderExternalImageApi

This trait is used as a bridge between the different GL clients in Servo that handles WebRender ExternalImages and the WebRender ExternalImageHandler API. This trait is used to notify lock/unlock messages and get the required info that WR needs.

WebViewTrait

A trait that exposes the embedding layer’s WebView to the Servo renderer. This is to prevent a dependency cycle between the renderer and the embedding layer.