Enum compositing_traits::constellation_msg::ConstellationMsg

pub enum ConstellationMsg {

    Exit,
    GetBrowsingContext(PipelineId, IpcSender<Option<BrowsingContextId>>),
    GetPipeline(BrowsingContextId, IpcSender<Option<PipelineId>>),
    GetFocusTopLevelBrowsingContext(IpcSender<Option<TopLevelBrowsingContextId>>),
    IsReadyToSaveImage(HashMap<PipelineId, Epoch>),
    Keyboard(KeyboardEvent),
    AllowNavigationResponse(PipelineId, bool),
    LoadUrl(TopLevelBrowsingContextId, ServoUrl),
    ClearCache,
    TraverseHistory(TopLevelBrowsingContextId, TraversalDirection),
    WindowSize(TopLevelBrowsingContextId, WindowSizeData, WindowSizeType),
    TickAnimation(PipelineId, AnimationTickType),
    WebDriverCommand(WebDriverCommandMsg),
    Reload(TopLevelBrowsingContextId),
    LogEntry(Option<TopLevelBrowsingContextId>, Option<String>, LogEntry),
    NewWebView(ServoUrl, TopLevelBrowsingContextId),
    CloseWebView(TopLevelBrowsingContextId),
    SendError(Option<TopLevelBrowsingContextId>, String),
    FocusWebView(TopLevelBrowsingContextId),
    BlurWebView,
    ForwardEvent(PipelineId, CompositorEvent),
    SetCursor(Cursor),
    EnableProfiler(Duration, Duration),
    DisableProfiler,
    ExitFullScreen(TopLevelBrowsingContextId),
    MediaSessionAction(MediaSessionActionType),
    SetWebViewThrottled(TopLevelBrowsingContextId, bool),
    IMEDismissed,
    ReadyToPresent(TopLevelBrowsingContextId),
    Gamepad(GamepadEvent),

}

Messages to the constellation.

Variants

Exit

Exit the constellation.

GetBrowsingContext(PipelineId, IpcSender<Option<BrowsingContextId>>)

Request that the constellation send the BrowsingContextId corresponding to the document with the provided pipeline id

GetPipeline(BrowsingContextId, IpcSender<Option<PipelineId>>)

Request that the constellation send the current pipeline id for the provided browsing context id, over a provided channel.

GetFocusTopLevelBrowsingContext(IpcSender<Option<TopLevelBrowsingContextId>>)

Request that the constellation send the current focused top-level browsing context id, over a provided channel.

IsReadyToSaveImage(HashMap<PipelineId, Epoch>)

Query the constellation to see if the current compositor output is stable

Keyboard(KeyboardEvent)

Inform the constellation of a key event.

AllowNavigationResponse(PipelineId, bool)

Whether to allow script to navigate.

LoadUrl(TopLevelBrowsingContextId, ServoUrl)

Request to load a page.

ClearCache

Clear the network cache.

TraverseHistory(TopLevelBrowsingContextId, TraversalDirection)

Request to traverse the joint session history of the provided browsing context.

WindowSize(TopLevelBrowsingContextId, WindowSizeData, WindowSizeType)

Inform the constellation of a window being resized.

TickAnimation(PipelineId, AnimationTickType)

Requests that the constellation instruct layout to begin a new tick of the animation.

WebDriverCommand(WebDriverCommandMsg)

Dispatch a webdriver command

Reload(TopLevelBrowsingContextId)

Reload a top-level browsing context.

LogEntry(Option<TopLevelBrowsingContextId>, Option<String>, LogEntry)

A log entry, with the top-level browsing context id and thread name

NewWebView(ServoUrl, TopLevelBrowsingContextId)

Create a new top level browsing context.

CloseWebView(TopLevelBrowsingContextId)

Close a top level browsing context.

SendError(Option<TopLevelBrowsingContextId>, String)

Panic a top level browsing context.

FocusWebView(TopLevelBrowsingContextId)

Make a top-level browsing context focused.

BlurWebView

Make none of the top-level browsing contexts focused.

ForwardEvent(PipelineId, CompositorEvent)

Forward an event to the script task of the given pipeline.

SetCursor(Cursor)

Requesting a change to the onscreen cursor.

EnableProfiler(Duration, Duration)

Enable the sampling profiler, with a given sampling rate and max total sampling duration.

DisableProfiler

Disable the sampling profiler.

ExitFullScreen(TopLevelBrowsingContextId)

Request to exit from fullscreen mode

MediaSessionAction(MediaSessionActionType)

Media session action.

SetWebViewThrottled(TopLevelBrowsingContextId, bool)

Set whether to use less resources, by stopping animations and running timers at a heavily limited rate.

IMEDismissed

Virtual keyboard was dismissed

ReadyToPresent(TopLevelBrowsingContextId)

Compositing done, but external code needs to present.

Gamepad(GamepadEvent)

Gamepad state has changed