compositing_traits::display_list

Struct ScrollTree

pub struct ScrollTree {
    pub nodes: Vec<ScrollTreeNode>,
}

A tree of spatial nodes, which mirrors the spatial nodes in the WebRender display list, except these are used for scrolling in Paint so that new offsets can be sent to WebRender.

Fields
nodes: Vec<ScrollTreeNode>

A list of Paint-side scroll nodes that describe the tree of WebRender spatial nodes, used by Paint to scroll the contents of the display list.

Implementations

impl ScrollTree

pub fn add_scroll_tree_node( &mut self, parent: Option<ScrollTreeNodeId>, info: SpatialTreeNodeInfo, ) -> ScrollTreeNodeId

Add a scroll node to this ScrollTree returning the id of the new node.

pub fn update_mapping(&mut self, mapping: Vec<SpatialId>)

Once WebRender display list construction is complete for this ScrollTree, update the mapping of nodes to WebRender SpatialIds.

pub fn get_node_mut(&mut self, id: ScrollTreeNodeId) -> &mut ScrollTreeNode

Get a mutable reference to the node with the given index.

pub fn get_node(&self, id: ScrollTreeNodeId) -> &ScrollTreeNode

Get an immutable reference to the node with the given index.

pub fn webrender_id(&self, id: ScrollTreeNodeId) -> SpatialId

Get the WebRender SpatialId for the given [ScrollNodeId]. This will panic if [ScrollTree::build_display_list] has not been called yet.

pub fn scroll_node_or_ancestor_inner( &mut self, scroll_node_id: ScrollTreeNodeId, scroll_location: ScrollLocation, context: ScrollType, ) -> Option<(ExternalScrollId, LayoutVector2D)>

fn node_with_external_scroll_node_id( &self, external_id: ExternalScrollId, ) -> Option<ScrollTreeNodeId>

pub fn scroll_node_or_ancestor( &mut self, external_id: ExternalScrollId, scroll_location: ScrollLocation, context: ScrollType, ) -> Option<(ExternalScrollId, LayoutVector2D)>

Scroll the scroll node with the given ExternalScrollId on this scroll tree. If the node cannot be scrolled, because it’s already scrolled to the maximum scroll extent, try to scroll an ancestor of this node. Returns the node scrolled and the new offset if a scroll was performed, otherwise returns None.

pub fn set_scroll_offset_for_node_with_external_scroll_id( &mut self, external_scroll_id: ExternalScrollId, offset: LayoutVector2D, context: ScrollType, ) -> Option<LayoutVector2D>

Given an ExternalScrollId and an offset, update the scroll offset of the scroll node with the given id.

pub fn set_all_scroll_offsets( &mut self, offsets: &FxHashMap<ExternalScrollId, LayoutVector2D>, )

Given a set of all scroll offsets coming from the Servo renderer, update all of the offsets for nodes that actually exist in this tree.

pub fn reset_all_scroll_offsets(&mut self)

Set the offsets of all scrolling nodes in this tree to 0.

pub fn scroll_offsets(&self) -> FxHashMap<ExternalScrollId, LayoutVector2D>

Collect all of the scroll offsets of the scrolling nodes of this tree into a HashMap which can be applied to another tree.

pub fn scroll_offset(&self, id: ExternalScrollId) -> Option<LayoutVector2D>

Get the scroll offset for the given ExternalScrollId or None if that node cannot be found in the tree.

pub fn cumulative_node_to_root_transform( &self, node_id: ScrollTreeNodeId, ) -> FastLayoutTransform

Find a transformation that can convert a point in the node coordinate system to a point in the root coordinate system.

pub fn cumulative_root_to_node_transform( &self, node_id: ScrollTreeNodeId, ) -> Option<FastLayoutTransform>

Find a transformation that can convert a point in the root coordinate system to a point in the coordinate system of the given node. This may be None if the cumulative transform is uninvertible.

pub fn cumulative_sticky_offsets( &self, node_id: ScrollTreeNodeId, ) -> LayoutVector2D

Find the cumulative offsets of sticky positioned boxes from the given node up to the root.

fn cumulative_node_transform( &self, node_id: ScrollTreeNodeId, ) -> ScrollTreeNodeTransformationCache

fn cumulative_node_transform_inner( &self, node: &ScrollTreeNode, ) -> ScrollTreeNodeTransformationCache

Traverse a scroll node to its root to calculate the transform.

fn invalidate_cached_transforms(&self)

fn external_scroll_id_for_scroll_tree_node( &self, id: ScrollTreeNodeId, ) -> Option<ExternalScrollId>

impl ScrollTree

Implementation of ScrollTree that is related to debugging.

fn nodes_in_adjacency_list(&self) -> Vec<Vec<ScrollTreeNodeId>>

fn debug_print_traversal( &self, print_tree: &mut PrintTree, current_id: ScrollTreeNodeId, adjacency_list: &[Vec<ScrollTreeNodeId>], )

pub fn debug_print(&self)

Print the ScrollTree. Particularly, we are printing the node in preorder traversal. The order of the nodes will depends of the index of a node in the ScrollTree which corresponds to the declarations of the nodes.