compositing_traits::display_list

Type Alias AdjacencyListForPrint

type AdjacencyListForPrint = Vec<Vec<ScrollTreeNodeId>>;

In order to pretty print the ScrollTree structure, we are converting the node list inside the tree to be a adjacency list. The adjacency list then is used for the ScrollTree::debug_print_traversal of the tree.

This preprocessing helps decouples print logic a lot from its construction.

Aliased Type

struct AdjacencyListForPrint { /* private fields */ }